back porting , Fix return type of ConceptService.getCountOfConceptReferenceTerms - TRUNK-2993
git-svn-id: http://svn.openmrs.org/openmrs/branches/1.9.x@25055 5bac5841-c719-aa4e-b3fe-cce5062f897a

+++ api/src/main/java/org/openmrs/api/db/hibernate/HibernateConceptDAO.java
@@ -1815,12 +1815,12 @@ public class HibernateConceptDAO implements ConceptDAO {
 	 * @see org.openmrs.api.db.ConceptDAO#getCountOfConceptReferenceTerms(java.lang.String, boolean)
 	 */
 	@Override
-	public Integer getCountOfConceptReferenceTerms(String query, ConceptSource conceptSource, boolean includeRetired)
+	public Long getCountOfConceptReferenceTerms(String query, ConceptSource conceptSource, boolean includeRetired)
 	        throws DAOException {
 		Criteria criteria = createConceptReferenceTermCriteria(query, conceptSource, includeRetired);
 		
 		criteria.setProjection(Projections.rowCount());
-		return (Integer) criteria.uniqueResult();
+		return (Long) criteria.uniqueResult();
 	}
 	
 	/**
